% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosswalk.R
\name{add_code_str}
\alias{add_code_str}
\title{convert a column from a list to a pipe-delimeted string for
easy viewing..}
\usage{
add_code_str(x, col, newcol)
}
\arguments{
\item{x}{a xwalk object}

\item{col}{the column that you want to make a string}

\item{newcol}{the new name of the column default= col_str}
}
\value{
a vector containing the string of concatenated occupational codes for each row
}
\description{
convert a column from a list to a pipe-delimeted string for
easy viewing..
}
