% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/action.R
\name{as_action}
\alias{as_action}
\alias{as_action.default}
\alias{as_action.event}
\title{Coerce tracking object to an action vector}
\usage{
as_action(event)

\method{as_action}{default}(event)

\method{as_action}{event}(event)
}
\arguments{
\item{event}{An event data.table. See \code{as_event()}.}
}
\value{
A vector of actions.
}
\description{
Coerce tracking object to an action vector
}
\examples{
data <- data.frame(period = 1, team = "home",
                   entity_start = "player1", entity_end = "player3",
                   time_start = 9.2, time_end = 11.1,
                   x_start = 33.2, x_end = 49.2,
                   y_start = 60.5, y_end = 67.1,
                   type = "pass", subtype = NA)
event <- as_event(data)
as_action(event)
}
