% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codingsystem.R
\name{as_codingsystem}
\alias{as_codingsystem}
\alias{as_codingsystem.data.frame}
\title{Create a coding system from a data frame}
\usage{
as_codingsystem(x, name = "", ...)

\method{as_codingsystem}{data.frame}(x, name = "", ...)
}
\arguments{
\item{x}{the data frame containing columns "code" and "title"}

\item{name}{coding system name}

\item{...}{additional parameters}
}
\value{
a codingsystem object.
}
\description{
Create a coding system from a data frame
}
