% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event.R
\name{as_event}
\alias{as_event}
\alias{as_event.default}
\alias{as_event.data.frame}
\title{Coerce lists, matrices, and data frames to event data.tables}
\usage{
as_event(data, ...)

\method{as_event}{default}(data, ...)

\method{as_event}{data.frame}(
  data,
  period = "period",
  team = "team",
  entity_start = "entity_start",
  entity_end = "entity_end",
  time_start = "time_start",
  time_end = "time_end",
  x_start = "x_start",
  x_end = "x_end",
  y_start = "y_start",
  y_end = "y_end",
  type = "type",
  subtype = "subtype",
  ...
)
}
\arguments{
\item{data}{A data frame, list, matrix, or other object that could be
coerced to a \code{data.table}.}

\item{...}{Unused, for extensibility.}

\item{period}{Name of column denoting the period.}

\item{team}{Name of column denoting the team.}

\item{entity_start}{Name of column denoting entity starting event.}

\item{entity_end}{Name of column denoting entity ending event.}

\item{time_start}{Name of column denoting time that event started.}

\item{time_end}{Name of column denoting time that event ended.}

\item{x_start}{Name of column denoting start position in x dimension.}

\item{x_end}{Name of column denoting end position in x dimension.}

\item{y_start}{Name of column denoting start position in y dimension.}

\item{y_end}{Name of column denoting end position in y dimension.}

\item{type}{Event type, e.g. "set piece".}

\item{subtype}{Event subtype, e.g. "kick off".}
}
\value{
An event data.table
}
\description{
Coerce lists, matrices, and data frames to event data.tables
}
\examples{
data <- data.frame(period = 1, team = "home",
                   entity_start = "player1", entity_end = "player3",
                   time_start = 9.2, time_end = 11.1,
                   x_start = 33.2, x_end = 49.2,
                   y_start = 60.5, y_end = 67.1,
                   type = "pass", subtype = NA)
as_event(data)
}
