% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/position.R
\name{as_position}
\alias{as_position}
\alias{as_position.default}
\alias{as_position.tracking}
\title{Coerce tracking object to a position vector}
\usage{
as_position(tracking, ball = TRUE, expand = FALSE)

\method{as_position}{default}(tracking, ball = TRUE, expand = FALSE)

\method{as_position}{tracking}(tracking, ball = TRUE, expand = FALSE)
}
\arguments{
\item{tracking}{A tracking data.table. See \code{as_tracking()}.}

\item{ball}{Include ball position? If \code{FALSE}, ball is not included.}

\item{expand}{Expand tracking data to include all (entity, time)
combinations.}
}
\value{
A vector of positions.
}
\description{
Coerce tracking object to a position vector
}
\examples{
data <- data.frame(period = 1, time = 0.04, team = "home",
                   entity = "player1", x = 3.62, y = 21.1)
tracking <- as_tracking(data)
as_position(tracking)
}
