% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOCcer.R
\name{codeJobHistory}
\alias{codeJobHistory}
\title{Auto-code job results with SOCcer using via the soccer API}
\usage{
codeJobHistory(title, task = "", industry = "", ..., n = 10)
}
\arguments{
\item{title}{The job title}

\item{task}{tasks performed on the job}

\item{industry}{industry (SIC 1987 code)}

\item{...}{(not used)}

\item{n}{the number of soc codes to return (default)}
}
\value{
a tibble consisting of the title/task/industry and the top n SOCcer results and scores
}
\description{
This codes one job at a time.  In order to code multiple jobs,
you can create a tibble (data frame) and use pmap_dfr to produce
a results tibble similar to the web-based version of SOCcer (https://soccer.nci.nih.gov)
}
\details{
Please use the web-based version for handling large jobs.
}
\examples{

soccer_results <- codeJobHistory("epidemiologist")
jobs <- tibble::tibble(title=c("chemist","farmer","data scientist"),
                       task=rep("",3),industry=rep("",3))
soccer_results_3 <- purrr::pmap_dfr(jobs,codeJobHistory,n=20)

}
