% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosswalk.R
\name{codes}
\alias{codes}
\title{get the codes for a crosswalk,}
\usage{
codes(x, code_column)
}
\arguments{
\item{x}{A crosswalk of class xwalk}

\item{code_column}{The column names for the desired codes.}
}
\description{
returns the all the codes in column code_column
}
