% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{codesAgree}
\alias{codesAgree}
\title{Check if codes agree with reviewer}
\usage{
codesAgree(codes, reviewer)
}
\arguments{
\item{codes}{codes to compare}

\item{reviewer}{reviewer's code -- "gold" standard}
}
\value{
TRUE if the codes are in the reviewer
}
\description{
Compares if a vector of codes is in a vector of reviewer
codes.
}
\details{
Particularly useful when combined with purrr::map_lgl
}
\examples{
x <- '11-1011'
codesAgree(x,c("11-1011","11-1021"))
codesAgree(x,c("13-1011","11-1021"))
}
