% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codingsystem.R
\name{select.codingsystem}
\alias{select.codingsystem}
\alias{filter.codingsystem}
\alias{mutate.codingsystem}
\alias{as_tibble.codingsystem}
\title{Use Coding system with dplyr}
\usage{
\method{select}{codingsystem}(.data, ...)

\method{filter}{codingsystem}(.data, ..., name = NULL, .by = NULL, .preserve = FALSE)

\method{mutate}{codingsystem}(.data, ...)

\method{as_tibble}{codingsystem}(x, ..., .rows = NULL, .name_repair = NULL, rownames = NULL)
}
\arguments{
\item{.data}{the coding system}

\item{...}{parts of the coding system}

\item{name}{name for the filtered coding system}

\item{.by}{passed to dplyr::filter}

\item{.preserve}{passed to dplyr::filter}

\item{x}{the coding system}

\item{.rows}{passed to dplyer::as_tibble}

\item{.name_repair}{passed to dplyer::as_tibble}

\item{rownames}{passed to dplyer::as_tibble}
}
\value{
a new codingsystem
}
\description{
These methods allow you to use the codingsystem like a tibble.
When using select, make sure you keep the code/title or else you
can break the functionality of the codingsystem.
}
