% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosswalk.R
\name{combine_crosswalks}
\alias{combine_crosswalks}
\title{Combine two crosswalks to produce a new crosswalk}
\usage{
combine_crosswalks(
  xw1,
  xw2,
  xw1_from,
  xw1_to,
  xw2_from,
  xw2_to,
  add_xwalk_string_column = TRUE
)
}
\arguments{
\item{xw1}{- crosswalk table 1}

\item{xw2}{- crosswalk table 2}

\item{xw1_from}{- the column in the xw1 table for the starting coding system}

\item{xw1_to}{- the column in the xw1 table for the intermediary coding system}

\item{xw2_from}{- the column in the xw2 table for the intermediary coding system}

\item{xw2_to}{- the column in the xw2 table for the final coding system}

\item{add_xwalk_string_column}{- boolean indicating whether a string equivalent should be added to the table}
}
\description{
Takes two concordance tables (xw1 and xw2), where xw1 go from coding system one
to an intermediary coding system, and xw2 goes from the intermediary coding system
to coding system two.  The goal is to make one table that goes from coding system 1 to
coding system 2.
}
\examples{
noc_isco <- readr::read_csv("https://danielruss.github.io/codingsystems/noc2011_isco2008.csv")
isco_soc <- readr::read_csv("https://danielruss.github.io/codingsystems/isco2008_soc2010.csv")
combine_crosswalks(noc_isco,isco_soc,noc2011_code,isco2008_code,isco2008_code,soc2010_code)

}
