% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codingsystem.R
\name{convert_column_type}
\alias{convert_column_type}
\alias{convert_column_type.codingsystem}
\title{Convert a column to a specified type}
\usage{
convert_column_type(x, col, type)

\method{convert_column_type}{codingsystem}(x, col, type)
}
\arguments{
\item{x}{A \code{codingsystem} object.}

\item{col}{An unquoted column name in \code{x$table} to convert.
A quoted column name is also accepted. To supply a name stored in a
variable, use \code{!!}, for example \code{!!column_name}.}

\item{type}{A character string specifying the target type: one of
\code{"integer"}, \code{"character"}, \code{"double"}, or
\code{"logical"}. Unambiguous abbreviations are accepted.}
}
\value{
A \code{codingsystem} object with the specified column in
  \code{table} converted to the requested type.
}
\description{
Convert a column in \code{x$table} to a base R type and return the
updated coding system.
}
\examples{
cs <- codingsystem(data.frame(code = "1", title = "Example", Level = "2"))
convert_column_type(cs, Level, "integer")
}
