% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml.R
\name{createMultiHotEncoder}
\alias{createMultiHotEncoder}
\title{creates a multihot encoder from a list of labels}
\usage{
createMultiHotEncoder(allLabels)
}
\arguments{
\item{allLabels}{The complete set of labels}
}
\value{
a function that preforms multihot encoding
}
\description{
creates a multihot encoder from a list of labels
}
