% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosswalk.R
\name{crosswalk.1}
\alias{crosswalk.1}
\title{looks up all corresponding codes from a crosswalk}
\usage{
crosswalk.1(
  codes,
  xwalk,
  from_column,
  to_column,
  add_xwalk_string_column = FALSE,
  unnest = FALSE
)
}
\arguments{
\item{codes}{the vector of codes that will be crosswalked}

\item{xwalk}{the concordance table.}

\item{from_column}{the column in the xwalk table for the codes of the starting coding system, if xwalk is of
class xwalk, and is unidirectional then this parameter is not needed.  If the xwalk if bidirectional,
then the parameter is optional. The default is xwalk$codes1.  If xwalk is not of class xwalk, then this parameter
is required}

\item{to_column}{the column in the xwalk table for the codes of the results coding system, if xwalk is of
class xwalk, and is unidirectional then this parameter is not needed.  If the xwalk if bidirectional,
then the parameter is optional. The default is xwalk$codes2.  If xwalk is not of class xwalk, then this parameter
is required}

\item{add_xwalk_string_column}{boolean indicating whether a string equivalent should be added to the table}

\item{unnest}{boolean indicating that the data should have multiple lines if there are multiple resulting codes.
This is not recommended, because the one input code could be repeated multiple times.}
}
\description{
Takes a vector of codes and concordance table (crosswalk) and converts from one
coding systems to the next.
}
