% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codingsystem.R
\name{dim.codingsystem}
\alias{dim.codingsystem}
\title{Dimensions of an Object}
\usage{
\method{dim}{codingsystem}(x)
}
\arguments{
\item{x}{an \R object, for example a matrix, array or data frame.}
}
\value{
For an array (and hence in particular, for a matrix) \code{dim} retrieves
  the \code{dim} attribute of the object.  It is \code{NULL} or a vector
  of mode \code{\link[base]{integer}}.

  The replacement method changes the \code{"dim"} attribute (provided the
  new value is compatible) and removes any \code{"dimnames"} \emph{and}
  \code{"names"} attributes.
}
\description{
Retrieve or set the dimension of an object.
}
\details{
The functions \code{dim} and \code{dim<-} are \link[base]{internal generic}
  \link[base]{primitive} functions.

  \code{dim} has a method for \code{\link[base]{data.frame}}s, which returns
  the lengths of the \code{row.names} attribute of \code{x} and
  of \code{x} (as the numbers of rows and columns respectively).
}
\examples{
x <- 1:12 ; dim(x) <- c(3,4)
x

# simple versions of nrow and ncol could be defined as follows
nrow0 <- function(x) dim(x)[1]
ncol0 <- function(x) dim(x)[2]
}
\references{
{\if{html}{\out{<span id="reference+dim.Rd+R+3ABecker+2BChambers+2BWilks+3A1988"></span>}}Becker RA, Chambers JM, Wilks AR (1988).
\emph{The New S Language}.
Chapman and Hall/CRC, London.
ISBN 053409192X.}
}
\seealso{
\code{\link[base]{ncol}}, \code{\link[base]{nrow}} and \code{\link[base]{dimnames}}.
}
