% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{displacement}
\alias{displacement}
\alias{displacement.position}
\title{Calculate displacement between positions}
\usage{
displacement(p, from = NULL, to = NULL)

\method{displacement}{position}(p, from = NULL, to = NULL)
}
\arguments{
\item{p}{A position vector.}

\item{from}{A vector of entity names, e.g. c("player1", "player2").}

\item{to}{A vector of entity names, e.g. c("player1", "player2").}
}
\value{
If \code{from = NULL} and \code{to = NULL}, a list. If \code{from} or \code{to} is an
an entity, a matrix with dimension \code{length(from)} * \code{length(to)}.
}
\description{
Calculate displacement between positions
}
\examples{
p <- position(entity = c("Havertz", "Mount"), team = "Chelsea",
              time = 10, x = c(15, 10), y = c(60, 50))
displacement(p, from = "Havertz", to = "Mount")
}
\seealso{
\code{\link[=distance]{distance()}}
}
