% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{distance}
\alias{distance}
\alias{distance.position}
\title{Calculate distance between positions}
\usage{
distance(p, from = NULL, to = NULL)

\method{distance}{position}(p, from = NULL, to = NULL)
}
\arguments{
\item{p}{A position vector.}

\item{from}{A vector of entity names, e.g. c("player1", "player2").}

\item{to}{A vector of entity names, e.g. c("player1", "player2").}
}
\value{
If \code{from = NULL} and \code{to = NULL}, a list. If \code{from} or \code{to} is an
an entity, a matrix with dimension \code{length(from)} * \code{length(to)}.
}
\description{
Calculate distance between positions
}
\seealso{
\code{\link[=displacement]{displacement()}}
}
