% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{expected_position}
\alias{expected_position}
\title{Calculate expected position of an entity}
\usage{
expected_position(p)
}
\arguments{
\item{p}{A position vector.}
}
\value{
An array with \code{length(p)} rows and columns "x" and "y".
}
\description{
Calculate expected position of an entity
}
