% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soc_2010.R
\name{extend_standard_soc1980_codes}
\alias{extend_standard_soc1980_codes}
\title{Extended SOC 1980 codes}
\usage{
extend_standard_soc1980_codes(codes)
}
\arguments{
\item{codes}{The codes we are extending}
}
\value{
extended codes.
}
\description{
Takes valid 1980 standardized codes (the ones in the book) and extends
them so that unit codes are always the most detailed (even if it is exactly
the same as the parent code.)
}
