% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codingsystem.R
\name{get_codes}
\alias{get_codes}
\title{Get a list of codes from a coding system}
\usage{
get_codes(.codingsystem)
}
\arguments{
\item{.codingsystem}{either a codingsystem or a tibble that has a a column
named "code".}
}
\value{
a vector of codes
}
\description{
Get a list of codes from a coding system
}
