% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_entity}
\alias{get_entity}
\title{Get entities from a position object}
\usage{
get_entity(x, team)
}
\arguments{
\item{x}{A position object.}

\item{team}{Character vector. Name of a team in \code{x}.}
}
\value{
A character vector of entities.
}
\description{
Get entities from a position object
}
\examples{
x <- position(entity = "Mount", team = "Chelsea",
              time = 1, x = 5, y = 6)
get_entity(x)
}
