% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_location}
\alias{get_location}
\title{Get location from a position object}
\usage{
get_location(x, entity)
}
\arguments{
\item{x}{A position object.}

\item{entity}{Character vector denoting entities in \code{x}.}
}
\value{
A matrix of locations with columns "x" and "y".
}
\description{
Get location from a position object
}
\examples{
x <- position(entity = "Mount", team = "Chelsea",
              time = 1, x = 5, y = 6)
get_location(x)
}
