% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_team}
\alias{get_team}
\title{Get team from a position object}
\usage{
get_team(x)
}
\arguments{
\item{x}{A position object.}
}
\value{
A character vector of teams.
}
\description{
Get team from a position object
}
\examples{
x <- position(entity = "Mount", team = "Chelsea",
              time = 1, x = 5, y = 6)
get_team(x)
}
