% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_time}
\alias{get_time}
\title{Get time from a position object}
\usage{
get_time(x)
}
\arguments{
\item{x}{A position object.}
}
\value{
A numeric vector of times.
}
\description{
Get time from a position object
}
\examples{
x <- position(entity = "Mount", team = "Chelsea",
              time = 1, x = 5, y = 6)
get_time(x)
}
