% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{influence_radius}
\alias{influence_radius}
\title{Calculate influence radius of a player}
\usage{
influence_radius(p)
}
\arguments{
\item{p}{A position vector.}
}
\value{
A numeric vector of \code{length(p)}
}
\description{
This function calculates the influence radius of player(s) given their
distance to the ball.
}
