% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/action.R
\name{is_action}
\alias{is_action}
\title{Test if the object is an action vector}
\usage{
is_action(x)
}
\arguments{
\item{x}{An object.}
}
\value{
\code{TRUE} if the object inherits from the \code{action} class.
}
\description{
This function returns \code{TRUE} for action vectors and \code{FALSE} for all
other objects.
}
