% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event.R
\name{is_event}
\alias{is_event}
\title{Test if the object is an event data.table}
\usage{
is_event(data)
}
\arguments{
\item{data}{An object.}
}
\value{
\code{TRUE} if the object inherits both \code{event} and
\code{data.table} classes.
}
\description{
This function returns \code{TRUE} for event data.tables and \code{FALSE} for all
other objects.
}
