% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/position.R
\name{is_position}
\alias{is_position}
\title{Test if the object is a position vector}
\usage{
is_position(x)
}
\arguments{
\item{x}{An object.}
}
\value{
\code{TRUE} if the object inherits from the \code{position} class.
}
\description{
This function returns \code{TRUE} for position vectors and \code{FALSE} for all
other objects.
}
