% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracking.R
\name{is_tracking}
\alias{is_tracking}
\title{Test if the object is a tracking data.table}
\usage{
is_tracking(data)
}
\arguments{
\item{data}{An object.}
}
\value{
\code{TRUE} if the object inherits both \code{tracking} and
\code{data.table} classes.
}
\description{
This function returns \code{TRUE} for tracking data.tables and \code{FALSE} for all
other objects.
}
