% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codingsystem.R
\name{is_url}
\alias{is_url}
\title{Check if a value is a url by looking
for the http(s)://
.Works with vectors...}
\usage{
is_url(x)
}
\arguments{
\item{x}{String to check}
}
\value{
logical vector TRUE if the x is a url False otherwise
}
\description{
Check if a value is a url by looking
for the http(s)://
.Works with vectors...
}
