% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codingsystem.R
\name{is_valid}
\alias{is_valid}
\title{Check if a set of codes are valid for a coding system}
\usage{
is_valid(code, system)
}
\arguments{
\item{code}{vector of codes to check}

\item{system}{the coding system}
}
\value{
boolean vector corresponding to whether the codes are in the coding system
}
\description{
Check if a set of codes are valid for a coding system
}
