% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soc_2010.R
\name{is_valid_4digit_noc2011}
\alias{is_valid_4digit_noc2011}
\title{Is a code valid 4-digit noc 2011 code?}
\usage{
is_valid_4digit_noc2011(code)
}
\arguments{
\item{code}{a character vector of noc 2011 codes}
}
\value{
a boolean vector
}
\description{
checks to see if a noc 2011 4 digit code is valid
}
