% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soc_2010.R
\name{is_valid_6d_soc2010}
\alias{is_valid_6d_soc2010}
\title{is a soc 2010 code valid?}
\usage{
is_valid_6d_soc2010(code)
}
\arguments{
\item{code}{a character vector of soc 2010 codes}
}
\value{
a boolean vector
}
\description{
checks to see if a soc 2010 6 digit code is valid
}
