% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soc_2010.R
\name{is_valid_6digit_soc2010}
\alias{is_valid_6digit_soc2010}
\title{Is a code valid 6-digit soc 2010 code?}
\usage{
is_valid_6digit_soc2010(code)
}
\arguments{
\item{code}{a character vector of soc 2010 codes}
}
\value{
a boolean vector
}
\description{
Is a code valid 6-digit soc 2010 code?
}
