% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soc_2010.R
\name{is_valid_soc1980}
\alias{is_valid_soc1980}
\title{Is a code valid soc 1980 code?}
\usage{
is_valid_soc1980(code)
}
\arguments{
\item{code}{a character vector of soc1980 codes}
}
\value{
a boolean vector
}
\description{
checks to see if a code is a valid standard soc1980 code
}
\seealso{
\code{\link{standardize_soc1980_codes}}
}
