% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{job_desc}
\alias{job_desc}
\title{Job Descriptions}
\format{
A data frame with 50 rows and 3 variables:
\describe{
\item{job_id}{unique identifier}
\item{title}{job title}
\item{description}{job descriptions}
}
}
\usage{
job_desc
}
\description{
A dataset containing 50 random job descriptions, titles, and IDs.
}
\keyword{datasets}
