% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codingsystem.R
\name{level}
\alias{level}
\alias{level.codingsystem}
\title{Get the code Level}
\usage{
level(data, codes)

\method{level}{codingsystem}(data, codes)
}
\arguments{
\item{data}{- a codingsystem}

\item{codes}{- a vector of codes to check}
}
\value{
a vector of Levels
}
\description{
Gets the levels for a vector of codes from a codingsystem
The type returned depends on the data.
}
\examples{
level(soc1980_all,"99-99") # "division"
level(soc2010_all,c("11-1011","11-2010")) # c(6,5)
}
