% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codingsystem.R
\name{load_codingsystem}
\alias{load_codingsystem}
\title{Load a coding system from a url or a path}
\usage{
load_codingsystem(url, name)
}
\arguments{
\item{url}{a url or a path to the coding system data}

\item{name}{the name for the codingsystem}
}
\value{
a codingsystem with data from the usl
}
\description{
Load a coding system from a url or a path
}
