% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOCAssign.R
\name{load_socassign_db}
\alias{load_socassign_db}
\title{load_socassign_db}
\usage{
load_socassign_db(fname, addSrc = FALSE)
}
\arguments{
\item{fname}{the SOCAssign db file}

\item{addSrc}{should I add the file name as an src column}
}
\value{
tibble with coder results
}
\description{
load the data from a SOCAssign SQLite database
}
