% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codingsystem.R
\name{lookup_code}
\alias{lookup_code}
\title{Look up code}
\usage{
lookup_code(x, system)
}
\arguments{
\item{x}{list of codes to lookup}

\item{system}{the coding system}
}
\value{
a vector of titles for the codes
}
\description{
Look up code
}
\examples{
soc2010 <- codingsystem(soc2010_6digit$soc_code,soc2010_6digit$title,"US SOC 2010 6 digit")
lookup_code( c("11-2031","Fred","11-3031") , soc2010)
}
