% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosswalk.R
\name{make_code_str}
\alias{make_code_str}
\title{convert a list column of codes to vector of string for display}
\usage{
make_code_str(x)
}
\arguments{
\item{x}{codes column}
}
\value{
a vector a string concatenating all the codes
}
\description{
convert a list column of codes to vector of string for display
}
\examples{
df <- tibble::tibble(soc2010_codes = list(c("11-1011","11-1021"),c("11-1000")))
df <- dplyr::mutate(df,code_str=make_code_str(soc2010_codes))
}
