% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosswalk.R
\name{make_combined_crosswalk}
\alias{make_combined_crosswalk}
\title{make a combined crosswalk}
\usage{
make_combined_crosswalk(xw1, xw2, xw1_from, xw1_to, xw2_from, xw2_to)
}
\description{
Takes two crosswalks and combines them...
}
