% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{metrica_event}
\alias{metrica_event}
\title{Sample event data from Metrica Sports}
\format{
A data frame with 892 rows and 12 variables:
\describe{
\item{team}{Name of a team}
\item{type}{Event type}
\item{subtype}{Event subtype}
\item{period}{Period of the match}
\item{time_start}{Time at start of event}
\item{time_end}{Time at end of event}
\item{entity_start}{Entity starting the event}
\item{entity_end}{Entity ending the event}
\item{x_start}{X position at start of event}
\item{x_end}{X position at end of event}
\item{y_start}{Y position at start of event}
\item{y_end}{Y position at end of event}
}
}
\source{
{https://github.com/metrica-sports/sample-data}
}
\usage{
metrica_event
}
\description{
Event data for sample game 1.
}
\keyword{datasets}
