% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{metrica_tracking}
\alias{metrica_tracking}
\title{Sample tracking data from Metrica Sports}
\format{
A data frame with 1,639,165 rows and 6 variables:
\describe{
\item{period}{Period of the match}
\item{time}{Time of match, measured in seconds}
\item{entity}{Entity, i.e. a player or the ball}
\item{x}{X position of the entity}
\item{y}{Y position of the entity}
\item{team}{Name of a team. For the ball entity, the team is "ball".}
}
}
\source{
{https://github.com/metrica-sports/sample-data}
}
\usage{
metrica_tracking
}
\description{
Tracking data for sample game 1.
}
\keyword{datasets}
