% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codingsystem.R
\name{name}
\alias{name}
\title{Returns the user assigned name of the coding system}
\usage{
name(system)
}
\arguments{
\item{system}{coding system}
}
\value{
the name of the coding system (may be blank)
}
\description{
Returns the user assigned name of the coding system
}
\examples{
soc2010 <- codingsystem(soc2010_6digit,name="US SOC 2010 6 digit")
name(soc2010)
}
