% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soc2010_6digit.R
\docType{data}
\name{noc2011_4digit}
\alias{noc2011_4digit}
\title{noc2011 4 digit classification system}
\format{
\describe{
\item{noc_code}{a 4 digit code formated like '0011', be careful must be a string not an integer}
\item{title}{a short definition of the code}
}
}
\source{
\url{https://danielruss.github.io/codingsystems/noc_2011_4d.csv}

\url{https://www.statcan.gc.ca/eng/subjects/standard/noc/2011/index}
}
\usage{
noc2011_4digit
}
\description{
Canadian 4 digit National Occupational Classification (NOC) 2011
}
\keyword{datasets}
