% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{noc2011_all}
\alias{noc2011_all}
\title{noc2011 4 digit classification system}
\format{
An object of class \code{codingsystem} of length 2.
}
\source{
\url{https://danielruss.github.io/codingsystems/noc011_all.csv}

\url{https://www.statcan.gc.ca/eng/subjects/standard/noc/2011/index}
}
\usage{
noc2011_all
}
\description{
Canadian  National Occupational Classification (NOC) 2011
}
\details{
The complete Canadian NOC 2011 system
}
\keyword{datasets}
