% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control.R
\name{pitch_control}
\alias{pitch_control}
\title{Model pitch control}
\usage{
pitch_control(tracking, grid = c(120, 80), cells = 200)
}
\arguments{
\item{tracking}{A tracking data.table. See \code{as_tracking()}.}

\item{grid}{A numeric vector giving grid dimensions, e.g. c(100, 50)}

\item{cells}{Number of cells at which to evaluate pitch control.}
}
\value{
A data.table with four columns giving the pitch control for each
(time, x, y) tuple in \code{tracking}.
}
\description{
Model pitch control
}
