% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/position.R
\name{position}
\alias{position}
\title{Build a position vector}
\usage{
position(entity, team, time = double(), x = double(), y = double())
}
\arguments{
\item{entity}{a vector of entities, e.g. "player1".}

\item{team}{a vector denoting team name, e.g. "Liverpool".}

\item{time}{a numeric vector of times.}

\item{x}{a numeric vector of x coordinates.}

\item{y}{a numeric vector of y coordinates.}
}
\value{
A position vector.
}
\description{
Build a position vector
}
\examples{
position(entity = "Mount", team = "Chelsea", time = 30.4, x = 70, y = 32)
}
