% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codingsystem.R
\name{print.codingsystem}
\alias{print.codingsystem}
\title{prints a codingsystem}
\usage{
\method{print}{codingsystem}(.codingsystem)
}
\arguments{
\item{.codingsystem}{}
}
\description{
prints a codingsystem
}
