% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rexport-dplyr.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{arrange}
\alias{filter}
\alias{mutate}
\alias{select}
\alias{count}
\alias{as_tibble}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr:arrange]{arrange()}}, \code{\link[dplyr:count]{count()}}, \code{\link[dplyr:filter]{filter()}}, \code{\link[dplyr:mutate]{mutate()}}, \code{\link[dplyr:select]{select()}}}

  \item{tibble}{\code{\link[tibble:as_tibble]{as_tibble()}}}
}}

