% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{soc1980_all}
\alias{soc1980_all}
\title{SOC 1980 complete classification system}
\format{
a coding system that consists of
\describe{
 \item{code}{The soc 1980 code}
 \item{title}{The title of the code}
}
}
\source{
\url{https://danielruss.github.io/codingsystems/soc1980_all.csv}
}
\usage{
soc1980_all
}
\description{
US SOC 1980 classification system
}
\details{
The complete US SOC 1980 system
}
\keyword{datasets}
