% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{soc1980_detailed}
\alias{soc1980_detailed}
\title{Detailed SOC 1980  classification system}
\format{
An object of class \code{codingsystem} of length 2.
}
\source{
\url{https://danielruss.github.io/codingsystems/soc1980_most_detailed.csv}
}
\usage{
soc1980_detailed
}
\description{
The US SOC 1980 classification system can have higher level (major or minor codes)
codes without any children. This data contains all the most detailed codes regardless
of the code level.
}
\keyword{datasets}
