% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{soc1980_extended}
\alias{soc1980_extended}
\title{Extended SOC 1980 complete classification system}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 995 rows and 8 columns.
}
\source{
\url{https://danielruss.github.io/codingsystems/soc_1980_extended.csv}
}
\usage{
soc1980_extended
}
\description{
The US SOC 1980 classification system can have higher level (major or minor codes)
codes without any children.  We extended the SOC 1980 classification system to
require all major codes (2-digit code) to have at least 1 minor code (3-digit code ),
and every minor codes to have at least 1 unit code (4-digit code).  The most
detailed code is now always a unit code.
}
\keyword{datasets}
