% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{soc2010_6digit}
\alias{soc2010_6digit}
\title{soc2010 6 digit classification system}
\format{
a coding system object
\describe{
  \item{code}{the 6-digit soc2010 code}
  \item{title}{the title of the 6-digit soc2010 code}
}
}
\source{
\url{https://danielruss.github.io/codingsystems/soc2010_all.csv}

\url{https://www.bls.gov/soc/2010/2010_major_groups.htm}
}
\usage{
soc2010_6digit
}
\description{
Downloaded by Daniel Russ
}
\details{
A codingsystem object with the soc2010 6-digit codes

The 6-digit soc2010 codes.
}
\keyword{datasets}
