% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{soc2018_all}
\alias{soc2018_all}
\title{Complete SOC 2018  classification system}
\format{
An object of class \code{codingsystem} of length 2.
}
\source{
\url{https://danielruss.github.io/codingsystems/soc2018_all.csv}
}
\usage{
soc2018_all
}
\description{
The complete US SOC 2018 classification system. This data contains all the codes
 regardless of the code level.
}
\keyword{datasets}
