% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOCcer-package.R
\docType{package}
\name{SOCcer-package}
\alias{SOCcer}
\alias{SOCcer-package}
\title{SOCcer: Query 'SOCcer API'}
\description{
Queries 'SOCcer API' and returns list in tidy format.
}
\author{
\strong{Maintainer}: Daniel Molitor \email{molitdj97@gmail.com}

}
\keyword{internal}
