% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soccer.R
\name{soccer}
\alias{soccer}
\title{Send and Retrive data from SOCcer API}
\usage{
soccer(
  job_title,
  job_desc = NULL,
  industry = NULL,
  n_results = 10,
  req_per_min = 5000,
  user_agent = "Daniel Molitor (dmolitor@ripl.org)",
  verbose = 0L,
  progress = FALSE,
  ...
)
}
\arguments{
\item{job_title}{Character string of job title.}

\item{job_desc}{Optional character string of job description.}

\item{industry}{Optional character string of industry code.}

\item{n_results}{Integer specifying the number of top matches to return.}

\item{req_per_min}{Integer specifying the max number of requests per minute.}

\item{user_agent}{Character string of user agent. Set to helpful value so
for API maintainer in case of serious problems.}

\item{verbose}{An integer between 0 and 3 specifying level of verbosity.}

\item{progress}{A boolean; whether to print progress across job
classifications. Will only work if the \code{progressr} package is installed.}

\item{...}{Additional arguments that will be mutated onto the returned data
frame.}
}
\value{
A list with two elements: \code{problems} which contains a tibble of all
requests that encountered an error and \code{valid} which contains a tibble of
all successfully validated results.
}
\description{
This function facilitates querying the SOCcer API in a vectorized manner.
Each argument must be either length 1 or n. This function facilitates request
throttling and verbose output.
}
\examples{
dat <- head(job_desc, 1)
soccer(job_title = dat$title,
       job_desc = dat$description,
       job_id = dat$job_id)

}
