% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/socR-package.R
\docType{package}
\name{socR-package}
\alias{socR}
\alias{socR-package}
\title{socR: Query 'SOCcer API'}
\description{
Queries 'SOCcer API' and returns list in tidy format.
}
\author{
\strong{Maintainer}: Daniel Molitor \email{molitdj97@gmail.com}

}
\keyword{internal}
