% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{speed}
\alias{speed}
\alias{speed.position}
\title{Calculate speed of an entity}
\usage{
speed(p)

\method{speed}{position}(p)
}
\arguments{
\item{p}{A position vector.}
}
\value{
A numeric vector of \code{length(p)}
}
\description{
Calculate speed of an entity
}
