% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{speed_ratio}
\alias{speed_ratio}
\title{Calculate speed of an entity as ratio of maximum speed}
\usage{
speed_ratio(p, max_speed = 13)
}
\arguments{
\item{p}{A position vector.}

\item{max_speed}{Assumed maximum speed.}
}
\value{
A numeric vector of \code{length(p)}
}
\description{
This function calculates the ratio between an entity's speed and a
fixed maximum speed
}
