% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml.R
\name{split_data}
\alias{split_data}
\title{Split data}
\usage{
split_data(x, pTrain = 0.9, pDev = 0.09, pTest = 0.01)
}
\arguments{
\item{x}{unused string data used to split the data}

\item{pTrain}{approximate percent of the training split}

\item{pDev}{approximate percent of the development split}

\item{pTest}{approximate percent of the test split}
}
\value{
a vector of factors (Train,Dev,Test) denoting the data split
}
\description{
A simple deterministic mechanism for splitting data into
training, development, and test data based on the MD5
hash of a unused string parameters.
}
\examples{
split_data(rownames(mtcars))
}
