% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soc_2010.R
\name{standardize_soc1980_codes}
\alias{standardize_soc1980_codes}
\title{Standardize US SOC 1980 codes}
\usage{
standardize_soc1980_codes(codes)
}
\arguments{
\item{codes}{vector of US SOC 1980 codes}
}
\value{
standardized US SOC 1980 codes
}
\description{
US SOC 1980 codes are often written in none stand form (e.g 4600 instead of 46-47).  This function
attempt to standardize some of the ways SOC 1980 codes are written.
}
\details{
the function trims leading and trailing zeros ("up to 2 trailing zero - 20 is a valid soc code)
}
\examples{
standardize_soc1980_codes(c("2000",'7600'))

}
