% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{theta}
\alias{theta}
\alias{theta.position}
\title{Calculate angle of an entity}
\usage{
theta(p)

\method{theta}{position}(p)
}
\arguments{
\item{p}{A position vector.}
}
\value{
A numeric vector of \code{length(p)}
}
\description{
This function calculates the angle of an entity's movement relative to the
x-axis
}
