% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codingsystem.R
\name{to_level}
\alias{to_level}
\title{to_level}
\usage{
to_level(codingsystem, level)
}
\arguments{
\item{codingsystem}{The coding system we are using}

\item{level}{The level in the coding system we want.  Should be a column name
in the codingsystem table.}
}
\value{
a function that converts a vector of codes from a lower level
 to a the level input.
}
\description{
A utility function for converting occupational codes to higher levels
in the hierarchy.
}
\examples{
to_soc2010_2d <- to_level(soc2010_all, soc2d)
to_soc2010_2d(c("11-1011","15-1110"))

}
