% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{to_list_column}
\alias{to_list_column}
\title{Create a list column from multiple columns}
\usage{
to_list_column(df, colname, ...)
}
\arguments{
\item{df}{the data frame you are modifying}

\item{colname}{the name of the new column}

\item{...}{the columns you are combining into a list column}
}
\value{
The original data frame with a new list column `colname` replacing
the columns given
}
\description{
This function replaces a set of input columns that you pass in with a list column
containing the values of input column on a row-by-row basis.
}
\examples{
df <- tibble::tibble(a_1=1:3,a_2=2:4,a_3=3:5,b=4:6) |> to_list_column(a,a_1,a_2,a_3)
}
