% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soc_2010.R
\name{valid_code}
\alias{valid_code}
\alias{is_valid_6digit_soc2010}
\alias{is_valid_soc1980}
\alias{is_most_detailed_soc1980}
\alias{is_valid_extended_soc1980}
\alias{is_most_detailed_extended_soc1980}
\title{Is valid code}
\usage{
valid_code(codeList)

is_valid_6digit_soc2010(code)

is_valid_soc1980(code)

is_most_detailed_soc1980(code)

is_valid_extended_soc1980(code)

is_most_detailed_extended_soc1980(code)
}
\arguments{
\item{codeList}{a vector of valid codes}

\item{code}{codes to compare}
}
\value{
valid_code returns a function. The functions (e.g. is_valid_soc2010)
take a code or a vector of codes and
returns a logic vector representing if the codes are valid.
}
\description{
check whether a code is valid
}
\details{
valid_code is a functional that create a function that check
if a vector of codes is valid

is_valid_4digit_soc1980, is_valid_6digit_soc2010 and is_valid_4digit_noc2011 were made using
valid_code functional.
}
\examples{
is_valid_toy <- valid_code(c("A","B","C"))
is_valid_toy(c("X","A","Z","B"))
}
\seealso{
[standardize_soc1980_codes()]
}
