% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{velocity}
\alias{velocity}
\alias{velocity.position}
\title{Calculate velocity of an entity}
\usage{
velocity(p)

\method{velocity}{position}(p)
}
\arguments{
\item{p}{A position vector.}
}
\value{
A numeric vector of \code{length(p)}.
}
\description{
Calculate velocity of an entity
}
