% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosswalk.R
\name{xwalk}
\alias{xwalk}
\title{xwalk class constructor}
\usage{
xwalk(dta, codes1, titles1, codes2, titles2, bidirectional = FALSE)
}
\description{
takes a data frame (the crosswalk) and which columns are the codes and titles
and create an xwalk object that can perform crosswalks...
}
