% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event.R
\name{as_event}
\alias{as_event}
\alias{as_event.default}
\alias{as_event.data.frame}
\title{Coerce lists, matrices, and data frames to event data.tables}
\usage{
as_event(data, ...)

\method{as_event}{default}(data, ...)

\method{as_event}{data.frame}(
  data,
  period = "period",
  team = "team",
  entity = c("entity_start", "entity_end"),
  time = c("time_start", "time_end"),
  x = c("x_start", "x_end"),
  y = c("y_start", "y_end"),
  type = "type",
  subtype = "subtype",
  ...
)
}
\arguments{
\item{data}{A data frame, list, matrix, or other object that could be
coerced to a \code{data.table}.}

\item{...}{Unused, for extensibility.}

\item{period}{A character vector denoting the period.}

\item{team}{A character vector denoting the team.}

\item{entity}{A character vector of variables denoting the entity
that started and ended an event.}

\item{time}{A character vector of variables denoting the times that
an event started and ended.}

\item{x}{A character vector of variables denoting the location that
an event started and ended in the y dimension.}

\item{type}{A character vector denoting the event type.}

\item{subtype}{A character vector denoting the event subtype.}
}
\value{
An event data.table
}
\description{
Coerce lists, matrices, and data frames to event data.tables
}
\examples{
data <- data.frame(period = 1, team = "home",
                   entity_start = "player1", entity_end = "player3",
                   time_start = 9.2, time_end = 11.1,
                   x_start = 33.2, x_end = 49.2,
                   y_start = 60.5, y_end = 67.1,
                   type = "pass", subtype = NA)
as_event(data)
}
