% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{bin_center}
\alias{bin_center}
\title{Bin Score and Assign center value}
\usage{
bin_center(score, n_bins)
}
\arguments{
\item{score}{The scores (values from 0-1) that are being binned}

\item{n_bins}{the number of bins the score}
}
\value{
the center of the score bin for all scores
}
\description{
The \code{bin_center()} function takes a vector of scores between 0-1 and
a number of bins an returns the center of the bin the score falls in.
}
