% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codingsystem.R
\name{codingsystem}
\alias{codingsystem}
\title{constructor create a coding system S3 class}
\usage{
codingsystem(codes, titles, name = "")
}
\arguments{
\item{codes}{vector of codes}

\item{titles}{vector of title}

\item{name}{coding system name}
}
\value{
the codingsystem object
}
\description{
constructor create a coding system S3 class
}
\examples{

codingsystem(soc2010_6digit$soc_code,soc2010_6digit$title,"US SOC 2010 6 digit")
}
