% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codingsystem.R
\name{codingsystem}
\alias{codingsystem}
\title{constructor create a coding system S3 class}
\usage{
codingsystem(codes, titles, name = "")
}
\arguments{
\item{codes}{vector of codes or a dataframe containing the columns "code" (with codes) and "title" (with titles)}

\item{titles}{vector of title}

\item{name}{coding system name}
}
\value{
the codingsystem object
}
\description{
constructor create a coding system S3 class
}
