% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codingsystem.R
\name{select.codingsystem}
\alias{select.codingsystem}
\alias{filter.codingsystem}
\alias{as_tibble.codingsystem}
\title{Use Coding system with dplyr}
\usage{
\method{select}{codingsystem}(.data, ...)

\method{filter}{codingsystem}(.data, ..., .by = NULL, .preserve = FALSE)

\method{as_tibble}{codingsystem}(x, ..., .rows = NULL, .name_repair = NULL, rownames = NULL)
}
\arguments{
\item{.data}{the coding system}

\item{...}{parts of the coding system}

\item{.by}{passed to dplyr::filter}

\item{.preserve}{passed to dplyr::filter}

\item{x}{the coding system}

\item{.rows}{passed to dplyer::as_tibble}

\item{.name_repair}{passed to dplyer::as_tibble}

\item{rownames}{passed to dplyer::as_tibble}
}
\value{
a tibble
}
\description{
Use Coding system with dplyr
}
