% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codingsystem.R
\name{select.codingsystem}
\alias{select.codingsystem}
\alias{filter.codingsystem}
\alias{as_tibble.codingsystem}
\title{Use Coding system with dplyr}
\usage{
\method{select}{codingsystem}(.codingsystem, ...)

\method{filter}{codingsystem}(.codingsystem, ...)

\method{as_tibble}{codingsystem}(.codingsystem, ...)
}
\arguments{
\item{.codingsystem}{the coding system}

\item{...}{parts of the coding system}
}
\value{
a tibble
}
\description{
Use Coding system with dplyr
}
