% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosswalk.R
\name{combine_crosswalks}
\alias{combine_crosswalks}
\title{Combine two crosswalks to produce a new crosswalk}
\usage{
combine_crosswalks(xw1, xw2)
}
\arguments{
\item{xw1}{- crosswalk 1, either an xwalk object or a data.frame}

\item{xw2}{- crosswalk 2,  either an xwalk object or a data.frame}
}
\description{
Takes two concordance tables (xw1 and xw2), where xw1 go from coding system one
to an intermediary coding system, and xw2 goes from the intermediary coding system
to coding system two.  The goal is to make one table that goes from coding system 1 to
coding system 2.
}
\examples{
# the noc_isco example has an extra column that confuses the parser,
# so I have to specify the parts or skip the last column.
noc_isco <- xwalk("https://danielruss.github.io/codingsystems/noc2011_isco2008.csv",
                   col_types = "cccc-")
isco_soc <- xwalk("https://danielruss.github.io/codingsystems/isco2008_soc2010.csv")
combine_crosswalks(noc_isco,isco_soc)

}
