% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosswalk.R
\name{crosswalk}
\alias{crosswalk}
\title{Use the concordance table (crosswalk) to convert from one
coding system to another.}
\usage{
crosswalk(codes, xwalk, from_column, to_column, invert = FALSE)
}
\arguments{
\item{codes}{the vector of codes that will be crosswalked}

\item{xwalk}{the concordance table.}

\item{invert}{by default the crosswalk goes from codes1 to
codes2 setting invert to TRUE make the crosswalk go from codes2
to codes1}
}
\value{
an unnamed list of codes in the resulting coding system
}
\description{
Use the concordance table (crosswalk) to convert from one
coding system to another.
}
