% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codingsystem.R
\name{format.codingsystem}
\alias{format.codingsystem}
\title{formats a codingsystem}
\usage{
\method{format}{codingsystem}(.codingsystem)
}
\arguments{
\item{.codingsystem}{}
}
\value{
a formatted character vector
}
\description{
formats a codingsystem
}
