% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codingsystem.R
\name{is.codingsystem}
\alias{is.codingsystem}
\title{checks if an object is a coding system}
\usage{
is.codingsystem(x)
}
\arguments{
\item{x}{object to test}
}
\description{
Is this object a coding system
}
\examples{
x <- codingsystem(soc2010_6digit$soc_code,soc2010_6digit$title,"US SOC 2010 6 digit")
is.codingsystem(x)
}
