% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup.R
\name{lookup_code}
\alias{lookup_code}
\title{Look up code}
\usage{
lookup_code(x, valid_codes, titles)
}
\arguments{
\item{x}{list of codes to lookup}

\item{valid_codes}{the set of legal codes, if an invalid codes is pass the function will return NA/}

\item{titles}{the set of title matching the codes.}
}
\value{
a vector of titles for the codes
}
\description{
Look up code
}
\examples{
x1 <- c("11-2031","Fred","11-3031")
lookup_code(x1,soc2010_6digit$soc_code,soc2010_6digit$title)
}
