% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{noc2011_4digit}
\alias{noc2011_4digit}
\title{noc2011 4 digit classification system}
\format{
An object of class \code{codingsystem} of length 2.
}
\source{
\url{https://danielruss.github.io/codingsystems/noc2011_all.csv}

\url{https://www.statcan.gc.ca/eng/subjects/standard/noc/2011/index}
}
\usage{
noc2011_4digit
}
\description{
Canadian 4 digit National Occupational Classification (NOC) 2011
}
\details{
A codingsystem object with the noc2011 4-digit codes
}
\keyword{datasets}
