% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codingsystem.R
\name{print.codingsystem}
\alias{print.codingsystem}
\title{prints a codingsystem}
\usage{
\method{print}{codingsystem}(x, ...)
}
\arguments{
\item{x}{- the codingsystem}

\item{...}{parameter for format, not currently used}
}
\description{
prints a codingsystem
}
