% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{soc1980_all}
\alias{soc1980_all}
\title{SOC 1980 complete classification system}
\format{
\describe{
\item{soc1980_code}{the n-digit soc 1980}
\item{title}{a short definition of the code}
}
}
\source{
\url{https://danielruss.github.io/codingsystems/soc1980_all.csv}
}
\usage{
soc1980_all
}
\description{
US SOC 1980 classification system
}
\keyword{datasets}
