% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soc2010_6digit.R
\docType{data}
\name{soc2010_6digit}
\alias{soc2010_6digit}
\title{soc2010 6 digit classification system}
\format{
\describe{
\item{soc_code}{a 6 digit code formated like '11-1011'}
\item{title}{a short definition of the code}
}
}
\source{
\url{https://danielruss.github.io/codingsystems/soc_2010_6digit.csv}

\url{https://www.bls.gov/soc/2010/2010_major_groups.htm}
}
\usage{
soc2010_6digit
}
\description{
Downloaded by Daniel Russ
}
\keyword{datasets}
