% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{soc2010_all}
\alias{soc2010_all}
\title{Complete SOC 2010  classification system}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 1425 rows and 9 columns.
}
\source{
\url{https://danielruss.github.io/codingsystems/soc2010_all.csv}
}
\usage{
soc2010_all
}
\description{
The complete US SOC 2010 classification system. This data contains all the codes
 regardless of the code level.
}
\keyword{datasets}
