% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/socR-package.R
\docType{package}
\name{socR-package}
\alias{socR}
\alias{socR-package}
\title{socR: Useful functions for working with occupation coding}
\description{
A set of functions that I find useful in my research into occupational coding and codes.
}
\author{
\strong{Maintainer}: Daniel Russ \email{druss@mail.nih.gov} (\href{https://orcid.org/0000-0003-4040-4416}{ORCID})

}
\keyword{internal}
