% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/socr-package.R
\docType{package}
\name{socr-package}
\alias{socr}
\alias{socr-package}
\title{socr: A framework for modeling soccer data}
\description{
Provides a standardized framework for modeling tracking and event data in soccer via coercion to 'position' and 'action' class objects.
}
\author{
\strong{Maintainer}: George Wood \email{georgebwood@gmail.com}

}
\keyword{internal}
