% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosswalk.R
\name{xwalk_entropy}
\alias{xwalk_entropy}
\title{Calculates the Shannon entropy for a Crosswalk}
\usage{
xwalk_entropy(x)
}
\arguments{
\item{x}{The crosswalk}
}
\value{
the entropy
}
\description{
The more potential codes that a crosswalk will allow an intial code to become, the higher
the entropy.  The entropy (H) is given by \deqn{H = \Sigma p log p} where p = 1/n and n is the number of potential codes
a single code can map to, natural logs are used in the calculation.
}
